+++ HISTORY.rst
@@ -4,7 +4,8 @@ Changelog
 2.5.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Do not delay on quality check
+  [vangheem]
 
 
 2.5.5 (2019-03-25)